requirements.txt
ansicon==1.89.0
anyio==4.12.1
asgiref==3.8.1
asttokens==3.0.0
black==25.11.0
blessed==1.23.0
CacheControl==0.14.4
catalogue==2.0.10
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.4
Choco==1.0.5
click==8.3.1
colorama==0.4.6
contexttimer==0.3.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.18.0
firebase_admin==7.1.0
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jinxed==1.3.0
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.3
matplotlib-inline==0.2.1
maven==0.0.6
msgpack==1.1.2
mypy_extensions==1.1.0
numpy==2.1.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.5
pathspec==0.12.1
pillow==11.2.1
pipenv==2025.0.2
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyreadline3==3.5.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2024.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==80.8.0
six==1.16.0
smmap==5.0.2
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.5.0
virtualenv==20.31.2
wcwidth==0.2.14
whitenoise==6.9.0
xlrd==2.0.2
